+++ source/dred_context.h
@@ -50,13 +50,13 @@ struct dred_context
 
 
     // The scale to apply for normal system DPI scaling. Usually this will be the same as UI scale, but not always.
-    double dpiScale;
+    float dpiScale;
 
     // The UI scale. All UI elements need to be scaled by this amount.
-    double uiScale;
+    float uiScale;
 
     // The scale to use for text editors.
-    double textEditorScale;
+    float textEditorScale;
 
 
     // Whether or not the application is running in silent mode.
@@ -248,10 +248,10 @@ void dred_toggle_line_numbers(dred_context* pDred);
 
 
 // Sets the scale of text editors.
-void dred_set_text_editor_scale(dred_context* pDred, double scale);
+void dred_set_text_editor_scale(dred_context* pDred, float scale);
 
 // Retrieves the scale of text editors.
-double dred_get_text_editor_scale(dred_context* pDred);
+float dred_get_text_editor_scale(dred_context* pDred);
 
 
 // Called when a tab is activated. This is where contextual GUI elements are shown or hidden.
